Remove pagination between posts
It tended to work in weird ways. If you looked at a presentation then it
would paginate between only presentations. If you looked at a post then
it would paginate between both posts and presentations. Until you got to
a presentation and then it would only do presentations.

I suspect something to do with how Jekyll inherits templates, but I do
not want to dig in to that right now.